ex01: used size_t insted of int for interation

+++ ex01/iter.hpp
@@ -6,7 +6,7 @@
 /*   By: cgodecke <cgodecke@student.42wolfsburg.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/29 16:42:56 by cgodecke          #+#    #+#             */
-/*   Updated: 2023/12/01 14:49:27 by cgodecke         ###   ########.fr       */
+/*   Updated: 2023/12/05 14:02:25 by cgodecke         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,9 @@
 # include<iostream>
 
 template<typename T_array, typename T_function>
-void iter(T_array *array, int length, T_function function)
+void iter(T_array *array, size_t length, T_function function)
 {
-	for (int i = 0; i < length; i++)
+	for (size_t i = 0; i < length; i++)
 	{
 		function(array[i]);
 	}
